PoC with all events in marcia

+++ src/bin/rsincrontab.rs
@@ -3,6 +3,7 @@ use rsincronlib::EVENT_TYPES;
 use std::{
     collections::HashMap,
     fs::{self, read_to_string, DirBuilder, File},
+    path::Path,
     process::Command,
 };
 use uuid::Uuid;
@@ -42,7 +43,11 @@ fn main() {
             ));
 
         let tmpfile_path = std::env::temp_dir().join(Uuid::new_v4().to_string());
-        File::create(&tmpfile_path).expect("couldn't open tmp file for writing: exiting");
+        if Path::new(&table_path).exists() {
+            fs::copy(&table_path, &tmpfile_path).expect("failed to open tmp file: exiting");
+        } else {
+            File::create(&tmpfile_path).expect("couldn't open tmp file for writing: exiting");
+        };
 
         let _exitstatus = Command::new(editor.clone())
             .arg(&tmpfile_path)
@@ -51,7 +56,12 @@ fn main() {
 
         let mut buf = String::new();
         for line in read_to_string(tmpfile_path).unwrap_or_default().lines() {
-            // TODO: implement all sort of checks plus error logging
+            if line.clone().chars().nth(0) == Some('#') {
+                buf.push_str(line);
+                buf.push('\n');
+                continue;
+            };
+
             let mut fields = line.split_whitespace();
 
             let Some(path) = fields.next() else {
@@ -77,6 +87,7 @@ fn main() {
 
             let table_entry = format!("{}\t{}\t{}", path, masks, command);
             buf.push_str(&table_entry);
+            buf.push('\n');
         }
 
         fs::write(&table_path, buf).expect(&format!(
